requirements.txt
duckdb==1.4.4
fredapi==0.5.2
lxml==6.0.2
pandas==3.0.1
pytest==9.0.2
python-dotenv==1.2.1
requests==2.32.5
yfinance==1.2.0
